Move SoupSieve obj to css_match SoupSieve object makes more sense in css_match as it is a wrapper around that class. Call CssParser on pattern before passing it in as CSSParser returns a hashable object that can be pickled.
