+++ soupsieve/__init__.py
@@ -27,6 +27,7 @@ SOFTWARE.
 """
 from .__meta__ import __version__, __version_info__  # noqa: F401
 from . import css_parser as cp
+from . import css_match as cm
 from .util import HTML, HTML5, XHTML, XML, deprecated
 
 __all__ = (
@@ -35,7 +36,7 @@ __all__ = (
     'comments', 'icomments', 'select', 'iselect', 'match', 'filter'
 )
 
-SoupSieve = cp.SoupSieve
+SoupSieve = cm.SoupSieve
 
 
 def compile(pattern, namespaces=None, mode=HTML5):  # noqa: A001
